create unziputil from mythcoreutil
