+++ mythtv/libs/libmythbase/mythcoreutil.h
@@ -9,14 +9,6 @@
 
  MBASE_PUBLIC  int64_t getDiskSpace(const QString &file_on_disk, int64_t &total, int64_t &used);
 
- MBASE_PUBLIC  bool extractZIP(QString zipFile, QString outDir);
-
- MBASE_PUBLIC  bool gzipFile(const QString &inFilename, const QString &zipFilename);
- MBASE_PUBLIC  bool gunzipFile(const QString &zipFilename, const QString &outFilename);
-
- MBASE_PUBLIC  QByteArray gzipCompress(const QByteArray &data);
- MBASE_PUBLIC  QByteArray gzipUncompress(const QByteArray &data);
-
  MBASE_PUBLIC  QString RemoteDownloadFile(const QString &url,
                                    const QString &storageGroup,
                                    const QString &filename = "");
