+++ mythtv/libs/libmythbase/test/test_unzip/test_unzip.cpp
@@ -18,7 +18,7 @@
  *   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
  */
 #include <iostream>
-#include "mythcoreutil.h"
+#include "unziputil.h"
 #include "test_unzip.h"
 #include <QTemporaryDir>
